+++ src/involvement/views/__init__.py
@@ -1,0 +1,13 @@
+# flake8: noqa
+from .admin_appoint import admin_appoint
+from .admin_approve_applicants import admin_approve_applicants
+from .admin_extend_deadline import admin_extend_deadline
+from .my_mandates import my_mandates
+from .open_positions import open_positions
+from .my_applications import my_applications
+from .view_position import view_position
+from .position_create_view import PositionCreateView
+from .position_edit_view import PositionEditView
+from .position_inspect_view import PositionInspectView
+from .role_create_view import RoleCreateView
+from .role_edit_view import RoleEditView
